updated driver to use Long.parseLong(...) to retrieve Integer results from BigQuery API.

+++ src/main/java/net/starschema/clouddb/jdbc/BQForwardOnlyResultSet.java
@@ -173,7 +173,7 @@ public class BQForwardOnlyResultSet implements java.sql.ResultSet {
                     return Boolean.parseBoolean(result);
                 }
                 if (Columntype.equals("INTEGER")) {
-                    return Integer.parseInt(result);
+                    return Long.parseLong(result);
                 }
                 if (Columntype.equals("TIMESTAMP")) {
                     long val = new BigDecimal(result).longValue() * 1000;
